Allow changing the name of an existing LB
Now we search LBRules on cloudstack both by name and by tags, this allow
us to find LBs even if a new name was set as a service annotation. If a
new name is found we'll delete the old LBRule on cloudstack and create
a new one with the new name reusing the IP address of the old LBRule.